+++ src/main/java/com/eeluwole/finance_api/policy/dto/PolicyResponse.java
@@ -3,6 +3,7 @@ package com.eeluwole.finance_api.policy.dto;
 import com.eeluwole.finance_api.policy.Policy.PolicyType;
 import com.eeluwole.finance_api.policy.Policy.PolicyStatus;
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -11,8 +12,8 @@ public class PolicyResponse {
     private Long id;
     private String policyNumber;
     private PolicyType type;
-    private Double coverageAmount;
-    private Double premiumAmount;
+    private BigDecimal coverageAmount;
+    private BigDecimal premiumAmount;
     private LocalDate startDate;
     private LocalDate endDate;
     private PolicyStatus status;
