+++ src/test/java/com/eeluwole/finance_api/beneficiary/BeneficiaryServiceTest.java
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -59,8 +60,8 @@ class BeneficiaryServiceTest {
         policy.setPolicyNumber("POL-001");
         policy.setClient(client);
         policy.setType(Policy.PolicyType.LIFE);
-        policy.setCoverageAmount(500000.0);
-        policy.setPremiumAmount(200.0);
+        policy.setCoverageAmount(BigDecimal.valueOf(500000.0));
+        policy.setPremiumAmount(BigDecimal.valueOf(200.0));
         policy.setStartDate(LocalDate.of(2024, 1, 1));
         policy.setEndDate(LocalDate.of(2025, 1, 1));
         policy.setStatus(Policy.PolicyStatus.ACTIVE);
